MAINT: create separate `form_factor` submodule
